requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
Babel==2.8.0
beautifulsoup4==4.8.2
bson==0.5.8
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
dataclasses==0.7
Django==2.2
django-autocomplete==0.9
django-autocomplete-light==3.5.1
django-autoslug-iplweb==1.9.4.dev0
django-bootstrap4==1.1.1
django-cities-light==3.5.0
django-compat==1.0.15
django-credit-cards==0.4.1
django-crispy-forms==1.9.0
django-jquery==3.1.0
django-model-utils==4.0.0
django-phonenumber-field==4.0.0
django-popup-forms==1.0.3
django-threadlocals==0.10
djongo==1.3.1
dnspython==1.16.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jsonfield==3.1.0
msgpack==0.6.2
phonenumbers==8.12.0
Pillow==7.0.0
pkg-resources==0.0.0
progressbar2==3.50.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pymongo==3.10.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2019.3
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.2.4
swapper==1.1.2.post1
Twisted==20.3.0
txaio==20.4.1
Unidecode==1.1.1
urllib3==1.25.8
zope.interface==5.1.0
